refactor: 컴포넌트 분리 및 클래스명 수정
