+++ src/page/MyStorePage/MyStorePage.module.scss
@@ -1,0 +1,62 @@
+.section {
+  width: 1131px;
+  margin: 0 auto;
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  height: 117px;
+  border-bottom: 2px solid #175c8e;
+
+  &__title {
+    height: 47px;
+    font-size: 36px;
+    font-weight: 500;
+    color: #175c8e;
+    font-family: "Noto Sans CJK KR", sans-serif;
+  }
+
+  &__btn-wrapper {
+    display: flex;
+    align-items: center;
+    gap: 16px;
+  }
+
+  &__btn-update {
+    width: 112px;
+    height: 48px;
+    border: 1px solid #175c8e;
+    font-size: 18px;
+    font-weight: 400;
+    color: #175c8e;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background-color: #ffffff;
+    font-family: "Noto Sans CJK KR", sans-serif;
+
+    &:hover {
+      cursor: pointer;
+    }
+  }
+
+  &__btn-add {
+    width: 112px;
+    height: 48px;
+    border: 1px solid #175c8e;
+    font-size: 18px;
+    font-weight: 400;
+    color: #ffffff;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    background: #175c8e;
+    font-family: "Noto Sans CJK KR", sans-serif;
+
+    &:hover {
+      cursor: pointer;
+    }
+  }
+}
