+++ src/page/MyStorePage/components/AddMenuModal/AddMenuModal.module.scss
@@ -1,0 +1,303 @@
+.modal-back {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgb(0 0 0 / 40%);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.modal {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  width: 810px;
+  height: 715px;
+  background-color: rgb(255 255 255);
+
+  &__header {
+    display: flex;
+    width: 100%;
+    height: 80px;
+    background-color: #175c8e;
+  }
+
+  &__title {
+    margin-left: 40px;
+    color: #ffffff;
+    display: flex;
+    align-items: center;
+    text-align: center;
+    font-size: 20px;
+    font-weight: 500;
+  }
+
+  &__add-menu {
+    display: flex;
+    font-size: 20px;
+  }
+
+  &__add-imgs {
+    width: 338px;
+  }
+
+  &__caption-title {
+    display: flex;
+  }
+
+  &__setting-title {
+    display: flex;
+    height: 27px;
+    margin-left: 40px;
+    margin-top: 25px;
+    font-weight: 700;
+    font-size: 18px;
+    align-items: center;
+    color: #175c8e;
+  }
+
+  &__caption {
+    display: flex;
+    height: 27px;
+    margin-top: 25px;
+    margin-left: 8px;
+    font-weight: 400;
+    font-size: 13px;
+    line-height: 19px;
+    align-items: center;
+    justify-content: center;
+    color: #a1a1a1;
+  }
+
+  &__add-new-img {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 266px;
+    height: 170px;
+    margin-top: 32px;
+    margin-left: 40px;
+    border: 1px dashed #858585;
+  }
+
+  &__add-new-img--icon {
+    width: 70px;
+    height: 70px;
+  }
+
+  &__add-new-img--caption {
+    display: flex;
+    margin-top: 16px;
+    font-weight: 400;
+    font-size: 16px;
+    align-items: center;
+    text-align: center;
+    color: #a1a1a1;
+  }
+
+  &__fence {
+    border-left: solid 1px #858585;
+    margin-top: 40px;
+    height: 520px;
+  }
+
+  &__info-setting {
+    width: 472px;
+  }
+
+  &__menu-name-input {
+    height: 27px;
+    width: 390px;
+    margin-top: 8px;
+    margin-left: 40px;
+    outline: none;
+    border: 0.5px solid #898a8d;
+    padding-left: 10px;
+
+    &::placeholder {
+      font-size: 13px;
+      font-weight: 400;
+      color: #898a8d;
+    }
+  }
+
+  &__size-wapper {
+    display: flex;
+    justify-content: space-between;
+    margin-right: 40px;
+  }
+
+  &__price-info-inputs {
+    display: flex;
+    width: 431px;
+    align-items: center;
+    margin-top: 8px;
+  }
+
+  &__size-input {
+    display: block;
+    margin-left: 40px;
+    height: 27px;
+    width: 175px;
+    margin-right: 8px;
+    outline: none;
+    border: 0.5px solid #898a8d;
+    padding-left: 10px;
+
+    &::placeholder {
+      font-size: 13px;
+      font-weight: 400;
+      color: #898a8d;
+    }
+  }
+
+  &__price-input {
+    display: block;
+    height: 27px;
+    width: 175px;
+    text-align: right;
+    outline: none;
+    border: 0.5px solid #898a8d;
+    padding-right: 7px;
+
+    &::placeholder {
+      font-size: 13px;
+      font-weight: 400;
+      color: #898a8d;
+    }
+  }
+
+  &__single-menu {
+    display: flex;
+    margin-top: 25px;
+    margin-left: 280px;
+    align-items: center;
+  }
+
+  &__single-menu--phrase {
+    width: 55px;
+    height: 27px;
+    font-weight: 400;
+    font-size: 14px;
+    display: flex;
+    align-items: center;
+    text-align: center;
+    color: #a1a1a1;
+  }
+
+  &__single-menu--icon {
+    margin-left: 8px;
+  }
+
+  &__cancle-icon {
+    margin-left: 15px;
+  }
+
+  &__add-price {
+    display: flex;
+    justify-items: center;
+  }
+
+  &__add-price-icon {
+    display: block;
+    width: 20px;
+    height: 20px;
+    margin-top: 14px;
+    margin-left: 40px;
+  }
+
+  &__add-price-word {
+    display: flex;
+    margin-top: 12px;
+    margin-left: 8px;
+    width: 63px;
+    height: 24px;
+    font-weight: 700;
+    font-size: 16px;
+    align-items: center;
+    color: #f7941e;
+  }
+
+  &__gear-icon {
+    margin-top: 28px;
+    margin-left: 190px;
+  }
+
+  &__menu-categoty-wapper {
+    display: flex;
+    flex-wrap: wrap;
+    margin: 0 40px;
+    margin-top: 8px;
+    gap: 8px;
+  }
+
+  &__menu-categoty-wapper--unit {
+    display: flex;
+    border: 0.5px solid #898a8d;
+    width: 190px;
+    height: 39px;
+    font-weight: 500;
+    font-size: 15px;
+    justify-content: center;
+    align-items: center;
+  }
+
+  &__menu-combo-input {
+    width: 390px;
+    height: 105px;
+    margin: 0 40px;
+    margin-top: 10px;
+    border: 0.5px solid #898a8d;
+    outline: none;
+    resize: none;
+
+    &::placeholder {
+      font-size: 13px;
+      font-weight: 400;
+      color: #898a8d;
+    }
+  }
+
+  &__btn-wrapper {
+    display: flex;
+    justify-content: flex-end;
+    gap: 32px;
+    margin-right: 40px;
+    margin-top: 20px;
+  }
+
+  &__delete-btn {
+    width: 114px;
+    height: 51px;
+    background: #858585;
+    border: none;
+    font-size: 18px;
+    font-weight: 400;
+    text-align: center;
+    color: #ffffff;
+
+    &:hover {
+      cursor: pointer;
+    }
+  }
+
+  &__submit-btn {
+    width: 114px;
+    height: 51px;
+    background: #175c8e;
+    border: none;
+    font-size: 18px;
+    font-weight: 400;
+    text-align: center;
+    color: #ffffff;
+
+    &:hover {
+      cursor: pointer;
+    }
+  }
+}
