+++ src/page/MyStorePage/components/Menus/Menus.module.scss
@@ -1,0 +1,94 @@
+.menu {
+  &__title-wrapper {
+    display: flex;
+    position: relative;
+  }
+
+  &__title {
+    margin-top: 24px;
+    margin-bottom: 10px;
+    height: 36px;
+    font-weight: 500;
+    font-size: 24px;
+    line-height: 36px;
+    display: flex;
+    align-items: center;
+    color: #175c8e;
+  }
+
+  &__event-menu-icon--mark {
+    margin-left: 8px;
+  }
+
+  &__wrapper {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+    align-items: center;
+  }
+
+  &__item {
+    box-sizing: border-box;
+    width: 368px;
+    height: 331px;
+    border: 1px solid #d2dae2;
+  }
+
+  &__empty-img {
+    position: relative;
+    margin: 0 auto;
+    margin-top: 39px;
+    width: 180px;
+    height: 180px;
+    border: 3px solid #175c8e;
+    border-radius: 100px;
+  }
+
+  &__empty-img--icon {
+    position: absolute;
+    top: 46.15px;
+    left: 62.32px;
+  }
+
+  &__empty-img--caption {
+    position: absolute;
+    color: #175c8e;
+    font-size: 14px;
+    font-weight: 500;
+    line-height: 21px;
+    text-align: center;
+    top: 113px;
+    left: 28px;
+    font-family: "Noto Sans CJK KR", sans-serif;
+  }
+
+  &__name {
+    display: block;
+    height: 30px;
+    font-size: 20px;
+    font-weight: 500;
+    line-height: 30px;
+    text-align: center;
+    color: #252525;
+    font-family: "Noto Sans CJK KR", sans-serif;
+    margin: 0 auto;
+    margin-top: 12px;
+  }
+
+  &__price {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 16px;
+    margin-top: 8px;
+  }
+
+  &__price--unit {
+    height: 24px;
+    font-size: 16px;
+    font-weight: 500;
+    text-align: left;
+    color: #f7941e;
+    font-family: "Noto Sans CJK KR", sans-serif;
+  }
+}
